+++ spead2/test/test_recv.py
@@ -653,6 +653,8 @@ class TestStream(object):
         assert_equal(0, stats.incomplete_heaps_evicted)
         assert_equal(0, stats.incomplete_heaps_flushed)
         assert_equal(1, stats.worker_blocked)
+        assert_equal(4, receiver.ringbuffer.capacity())
+        assert_equal(4, receiver.ringbuffer.size())
 
         receiver.stop()            # This unblocks all remaining heaps
         stats = receiver.stats
@@ -661,6 +663,8 @@ class TestStream(object):
         assert_equal(0, stats.incomplete_heaps_evicted)
         assert_equal(0, stats.incomplete_heaps_flushed)
         assert_equal(1, stats.worker_blocked)
+        assert_equal(4, receiver.ringbuffer.capacity())
+        assert_equal(4, receiver.ringbuffer.size())
 
     def test_no_stop_heap(self):
         """A heap containing a stop is not passed to the ring"""